+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Add Object.setPath({}, path, value)
 * Add JSON.dry and JSON.undry (circular & reviver support)
 * Add String#assign and String#assignments
+* Add Iterator#reset, to go back to the beginning
 
 ## 0.1.3 (2014-09-15)